% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execQuery.R
\name{execQuery}
\alias{execQuery}
\title{Load the result of an OMOP query.}
\usage{
execQuery(
  domain = NULL,
  query_name = NULL,
  input = NULL,
  where_clause = NULL,
  row_limit = NULL,
  row_offset = 0,
  db_connection = NULL,
  cdm_schema = "public",
  vocabulary_schema = " public"
)
}
\arguments{
\item{domain}{a character, the query domain (ex. 'care_site')}

\item{query_name}{the query name as it appears in the result of showQueries}

\item{input}{a vector of input parameters, in the same order as they appear in the text of the query. Information is available in showQueries() for each case.}

\item{where_clause}{an optional where clause that can be appended to the query (without the 'where' keyword)}

\item{row_limit}{maximum number of rows retrieved}

\item{db_connection}{a character, the name of the connection to the database. It must already exist in the R session.
If no db_connection is provided, the first connection found in the session will be used.}
}
\value{
the query result
}
\description{
Load the resuults of one of the preset queries (previously retrieved with showQueries)
}
