% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApp.R
\name{launchQueryLibrary}
\alias{launchQueryLibrary}
\title{Launch the SqlRender Developer Shiny app}
\usage{
launchQueryLibrary(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Should the app be launched in your default browser, or in a Shiny window.
Note: copying to clipboard will not work in a Shiny window.}
}
\description{
Launch the SqlRender Developer Shiny app
}
\details{
Launches a Shiny app that allows the user to select queries from the library and run these against a CDM
}
