% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showQueries.R
\name{showQueries}
\alias{showQueries}
\title{Show all available queries}
\usage{
showQueries(
  query_type = c("Data load", "General statistics"),
  domain = NULL,
  query_name = NULL
)
}
\arguments{
\item{query_type}{a character, one of  "Data load" or "General statistics". General statistics: generic statistics queries, same as the ones provided by the QueryLibrary OHDSI app.
Data load: queries that load data in the R session, preformatted for further processing.}

\item{domain}{a character, used to restrict the retrieved queries by domain (a list of all available domains is shown when invoking with domain = NULL)}

\item{query_name}{a character, restricts  the result to one specific query}
}
\value{
a list containing all the available queries with documentation
}
\description{
Return all the available queries explained and grouped by domain.
}
