% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execQuery.R
\name{execQuery}
\alias{execQuery}
\title{Load the result of an OMOP query in the remote session(s)}
\usage{
execQuery(
  domain = NULL,
  query_name = NULL,
  input = NULL,
  where_clause = NULL,
  row_limit = NULL,
  row_offset = 0,
  db_connection = NULL,
  cdm_schema = "public",
  vocabulary_schema = " public"
)
}
\arguments{
\item{domain}{a character, the query domain (ex. 'care_site')}

\item{query_name}{the query name as it appears in the result of dsqShowQueries}

\item{input}{a vector of input parameters, in the same order as they appear in the text of the query. Information is available in dsqShowQueries() for each case.}

\item{where_clause}{an optional where clause that can be appended to the query (without the 'where' keyword)}

\item{row_limit}{maximum number of row retrieved}

\item{symbol}{a character, the name of the data frame where the resultset will be loaded. It defaults to the query_name.}

\item{db_connections}{a vector, the name of the connection(s) to the database. They must exist in the remote session(s) - it can be created with datashield.assign.db_connection().
If no db_connection is provided, all the connections found in the renote session will be used. One more column called "database"
will be added to the query results and it will be populated with the respective connection names. The next argument (union)
will govern further behaviour.}
}
\value{
the query result
}
\description{
Load remotely one of the preset queries (previously retrieved with dsqShowQueries)
}
